refactor: update module options
